refactor : GameResultTest수정
