+++ src/main/java/domain/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package util;
+package domain;
 
 import java.util.Random;
