+++ src/test/java/util/RandomNumberGeneratorTest.java
@@ -1,5 +1,6 @@
 package util;
 
+import domain.RandomNumberGenerator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
